+++ packages/builders/phosphor-icons/src/routes/+page.svelte
@@ -2,12 +2,14 @@
 	import { Icon } from '@steeze-ui/svelte-icon'
 	import * as Icons from '$lib/index.js'
 
-	const themes = ['default', '24', '12']
+	const icons: any = Icons
+
+	const themes = Object.keys(icons[Object.keys(icons)[0]])
 
 	let theme = 'default'
 </script>
 
-<h1>@steeze-ui/octicons</h1>
+<h1>Your Custom Iconpack</h1>
 
 <label for="theme-select">Theme</label>
 <select id="theme-select" bind:value={theme}>
@@ -16,8 +18,8 @@
 	{/each}
 </select>
 <div>
-	{#each Object.keys(Icons) as Src}
-		<Icon src={Icons[Src]} size="24px" {theme} />
+	{#each Object.keys(icons) as Src}
+		<Icon src={icons[Src]} size="24px" {theme} class="blue" />
 	{/each}
 </div>
 
@@ -26,5 +28,9 @@
 		display: flex;
 		flex-wrap: wrap;
 		gap: 1rem;
+	}
+
+	:global(.blue) {
+		color: blue;
 	}
 </style>