+++ packages/builders/template-builder/CHANGELOG.md
@@ -1,1 +1,5 @@
-# @steeze-ui/template-icons
+# @steeze-ui/unnamed
+
+## 1.0.0
+
+- initial release